requirements.txt
agent-framework --pre
openai>=1.30.0
python-dotenv
requests
beautifulsoup4
pydantic
python-docx
dspy>=3.1,<4.0
tiktoken
streamlit